:sparkles: mqtt codec 代码优化。

+++ mica-mqtt-codec/src/main/java/net/dreamlu/iot/mqtt/codec/MqttMessageType.java
@@ -25,63 +25,63 @@ public enum MqttMessageType {
 	/**
 	 * 连接服务端
 	 */
-	CONNECT(1),
+	CONNECT((byte) 1),
 	/**
 	 * 确认连接请求
 	 */
-	CONNACK(2),
+	CONNACK((byte) 2),
 	/**
 	 * 发布消息
 	 */
-	PUBLISH(3),
+	PUBLISH((byte) 3),
 	/**
 	 * 发布确认
 	 */
-	PUBACK(4),
+	PUBACK((byte) 4),
 	/**
 	 * 发布收到（QoS 2，第一步）
 	 */
-	PUBREC(5),
+	PUBREC((byte) 5),
 	/**
 	 * 发布释放（QoS 2，第二步）
 	 */
-	PUBREL(6),
+	PUBREL((byte) 6),
 	/**
 	 * 发布完成（QoS 2，第三步）
 	 */
-	PUBCOMP(7),
+	PUBCOMP((byte) 7),
 	/**
 	 * 订阅主题
 	 */
-	SUBSCRIBE(8),
+	SUBSCRIBE((byte) 8),
 	/**
 	 * 订阅确认
 	 */
-	SUBACK(9),
+	SUBACK((byte) 9),
 	/**
 	 * 取消订阅
 	 */
-	UNSUBSCRIBE(10),
+	UNSUBSCRIBE((byte) 10),
 	/**
 	 * 取消订阅确认
 	 */
-	UNSUBACK(11),
+	UNSUBACK((byte) 11),
 	/**
 	 * 心跳请求
 	 */
-	PINGREQ(12),
+	PINGREQ((byte) 12),
 	/**
 	 * 心跳响应
 	 */
-	PINGRESP(13),
+	PINGRESP((byte) 13),
 	/**
 	 * 断开连接
 	 */
-	DISCONNECT(14),
+	DISCONNECT((byte) 14),
 	/**
 	 * 认证
 	 */
-	AUTH(15);
+	AUTH((byte) 15);
 
 	private static final MqttMessageType[] VALUES;
 
@@ -91,7 +91,7 @@ public enum MqttMessageType {
 		final MqttMessageType[] values = values();
 		VALUES = new MqttMessageType[values.length + 1];
 		for (MqttMessageType mqttMessageType : values) {
-			final int value = mqttMessageType.value;
+			final byte value = mqttMessageType.value;
 			if (VALUES[value] != null) {
 				throw new AssertionError("value already in use: " + value);
 			}
@@ -99,13 +99,13 @@ public enum MqttMessageType {
 		}
 	}
 
-	private final int value;
+	private final byte value;
 
-	MqttMessageType(int value) {
+	MqttMessageType(byte value) {
 		this.value = value;
 	}
 
-	public int value() {
+	public byte value() {
 		return value;
 	}
